refactor(elementAt): moved out of extended directory
BREAKING CHANGE: `elementAt` has been moved out of `extended` folder and under `operator` folder directly
