--- src/add/operator/extended/elementAt.ts
@@ -1,7 +1,0 @@
-import {Observable} from '../../../Observable';
-import {elementAt} from '../../../operator/extended/elementAt';
-import {KitchenSinkOperators} from '../../../Rx.KitchenSink';
-const observableProto = (<KitchenSinkOperators<any>>Observable.prototype);
-observableProto.elementAt = elementAt;
-
-export var _void: void;
